Output script
Powershell script running on the admin server, output the reservation inactivity in days via the multiple exports from the DHCP servers.
From the reservation name, if possible it tries to contact a Domain Controler to get the "LastLogonTimeStamp" value et determine when the computer was last seen on the network.